# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/suchet/smartFridge/stocks/StockActivity.java
@@ -23,7 +23,6 @@ import java.util.List;
 public class StockActivity extends AppCompatActivity {
     private ActivityStockBinding binding;
 
-
     private StockAdapter stockAdapter;
 
     @Override
@@ -37,6 +36,7 @@ public class StockActivity extends AppCompatActivity {
         binding.displayStock.setLayoutManager(new LinearLayoutManager(this));
         displayStock();
         GoToAddStockActivity();
+        GoToDeleteStockActivity();
         backToLanding();
     }
 
@@ -73,22 +73,32 @@ public class StockActivity extends AppCompatActivity {
             }
         });
     }
-
+    private void GoToDeleteStockActivity() {
+        binding.deleteFoodInStockButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                startActivity(DeleteStockActivity.DeleteStockIntentFactory(getApplicationContext()));
+            }
+        });
+    }
     public static void addFoodToStock(Context context, Food food) {
-        Log.d("STOCK", "Food 1 : appelle dans stockactivity " + food.getName());
         new Thread(() -> {
             SmartFridgeDatabase stockDatabase = SmartFridgeDatabase.getDatabase(context);
-            Log.d("STOCK", "Food 2 : rentre dans thread");
             String username = context.getSharedPreferences("user_session", Context.MODE_PRIVATE)
                     .getString("current_username", null);
-            Log.d("STOCK", "Food 3 : username ? " + username);
             if (username == null) return;
 
             User user = stockDatabase.userDAO().getUserByUsernameSync(username);
-            Log.d("STOCK", "Food 4 : bon user ? " + user.getUsername());
+            List<Food> list = stockDatabase.foodDAO().getFoodByUser(user.getId());
+            for(Food f : list){
+                if(f.getName().equals(food.getName())){
+                    f.setQuantity(f.getQuantity() + food.getQuantity());
+                    stockDatabase.foodDAO().update(f);
+                    return;
+                }
 
+            }
             food.setUserId(user.getId());
-            food.setDatePeremption(LocalDate.now().plusDays(1));
             stockDatabase.foodDAO().insert(food);
         }).start();
 
@@ -98,12 +108,22 @@ public class StockActivity extends AppCompatActivity {
     @Override
     protected void onResume() {
         super.onResume();
-        displayStock(); // refresh la liste à chaque fois qu’on revient sur l’activité
+        displayStock();
     }
 
     public static void deleteFoodToStock(Context context, Food food) {
-        SmartFridgeDatabase stockDatabase = SmartFridgeDatabase.getDatabase(context);
-        stockDatabase.foodDAO().delete(food);
+        new Thread(() -> {
+            SmartFridgeDatabase stockDatabase = SmartFridgeDatabase.getDatabase(context);
+            String username = context.getSharedPreferences("user_session", Context.MODE_PRIVATE)
+                    .getString("current_username", null);
+            if (username == null) return;
+
+            User user = stockDatabase.userDAO().getUserByUsernameSync(username);
+
+            food.setUserId(user.getId());
+            stockDatabase.foodDAO().delete(food);
+        }).start();
+
     }
     public static Intent StockIntentFactory(Context context) {
         Intent intent = new Intent(context, StockActivity.class);
@@ -114,19 +134,15 @@ public class StockActivity extends AppCompatActivity {
         List<Food> foodForTomorrow = new ArrayList<>();
         new Thread(() -> {
             SmartFridgeDatabase stockDatabase = SmartFridgeDatabase.getDatabase(context);
-            Log.d("STOCK", "Food 1 : rentre dans thread pour getFoodForTomorrow");
             String username = context.getSharedPreferences("user_session", Context.MODE_PRIVATE)
                     .getString("current_username", null);
-            Log.d("STOCK", "Food 2 : username ? " + username);
             if (username == null) return;
 
             User user = stockDatabase.userDAO().getUserByUsernameSync(username);
-            Log.d("STOCK", "Food 3 : bon user ? " + user.getUsername());
             List<Food> foodList = stockDatabase.foodDAO().getFoodByUser(user.getId());
             for (Food food : foodList) {
                 if(food.getDatePeremption() != null && food.getDatePeremption().isEqual(java.time.LocalDate.now().plusDays(1))){
                     foodForTomorrow.add(food);
-                    Log.d("STOCK", "Food 4 : ajout de " + food.getName() + " dans la liste pour demain" + food.getDatePeremption());
                 }
 
             }
